convert publish to pypi job as well
