+++ tests/version_test.py
@@ -1,13 +1,9 @@
 from __future__ import annotations
 
 import re
-from collections.abc import Callable
 from importlib.metadata import distribution
 from os import environ
 from pathlib import Path
-from typing import TYPE_CHECKING, Any
-
-import pytest
 
 import narwhals as nw
 from narwhals._utils import parse_version
@@ -39,20 +35,3 @@ def test_package_version() -> None:
         assert version_comp <= parse_version(pyproject_version)
     dist_version = distribution("narwhals").version
     assert version == dist_version
-
-
-def test_package_getattr() -> None:
-    pytest.importorskip("typing_extensions")
-    from typing_extensions import assert_type
-
-    ok = nw.__version__
-    assert_type(ok, str)
-    also_ok = nw.all
-    assert_type(also_ok, Callable[[], nw.Expr])
-
-    if TYPE_CHECKING:
-        bad = nw.not_real  # type: ignore[attr-defined]
-        assert_type(bad, Any)
-
-    with pytest.raises(AttributeError):
-        very_bad = nw.not_real  # type: ignore[attr-defined]  # noqa: F841
